requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
pyfiglet==1.0.4
python-dotenv==1.2.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
typing-inspection==0.4.2
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
